+++ Makefile.am
@@ -145,11 +145,22 @@ tcti_libtcti_device_la_SOURCES  = tcti/tcti_device.c tcti/commonchecks.c \
 tcti_libtcti_socket_la_CFLAGS   = $(AM_CFLAGS)
 tcti_libtcti_socket_la_LDFLAGS  = -Wl,--version-script=$(srcdir)/tcti/tcti_socket.map
 tcti_libtcti_socket_la_SOURCES  = tcti/platformcommand.c tcti/tcti_socket.c \
-                                  tcti/commonchecks.c common/sockets.c common/debug.c
+                                  tcti/commonchecks.c tcti/sockets.c common/debug.c
 
 test_tpmclient_tpmclient_CFLAGS   = $(AM_CFLAGS) -DNO_RM_TESTS
 test_tpmclient_tpmclient_LDADD    = $(libsapi) $(libtcti_socket) $(libtcti_device) $(libmarshal)
-test_tpmclient_tpmclient_SOURCES  = test/tpmclient/tpmclient.c $(COMMON_C) $(SAMPLE_C)
+test_tpmclient_tpmclient_SOURCES  = $(COMMON_C) \
+    test/tpmclient/CatSizedByteBuffer.c test/tpmclient/CopySizedBuffer.c \
+    test/tpmclient/CreateNullSession.c test/tpmclient/DecryptEncrypt.c \
+    test/tpmclient/Entity.c test/tpmclient/kdfa.c \
+    test/tpmclient/LoadExternalHMACKey.c test/tpmclient/sample.h \
+    test/tpmclient/SessionHmac.c test/tpmclient/SetLocality.c \
+    test/tpmclient/StartAuthSession.c test/tpmclient/syscontext.c \
+    test/tpmclient/syscontext.h test/tpmclient/tcti_util.c \
+    test/tpmclient/tcti_util.h test/tpmclient/TpmCalcPHash.c \
+    test/tpmclient/tpmclient.c test/tpmclient/tpmclient.h \
+    test/tpmclient/tpmclient_wo_rm.h test/tpmclient/TpmHandleToName.c \
+    test/tpmclient/TpmHash.c test/tpmclient/TpmHmac.c
 
 test_integration_libtest_utils_la_SOURCES = test/integration/test-options.c \
     test/integration/context-util.c
