+++ tcti/sockets.h
