+++ test/tpmclient/CopySizedBuffer.c
