+++ test/tpmclient/LoadExternalHMACKey.c
