+++ test/tpmclient/SessionHmac.c
