+++ test/tpmclient/SetLocality.c
