+++ test/tpmclient/TpmCalcPHash.c
