+++ test/tpmclient/TpmHash.c
