+++ test/tpmclient/kdfa.c
