+++ test/tpmclient/sample.h
@@ -34,10 +34,10 @@ extern "C" {
 
 
 #include "sapi/tss2_tpm2_types.h"
-#include "test/tpmclient/tpmclient.h"
+#include "tpmclient.h"
 #include <stdio.h>
 #include <stdlib.h>
-#include "common/syscontext.h"
+#include "syscontext.h"
 #include "common/debug.h"
 
 extern TSS2_TCTI_CONTEXT *resMgrTctiContext;
