+++ test/tpmclient/tcti_util.c
@@ -1,5 +1,5 @@
 #include "tcti_util.h"
-#include "debug.h"
+#include "common/debug.h"
 
 TSS2_RC InitDeviceTctiContext( const TCTI_DEVICE_CONF *driverConfig, TSS2_TCTI_CONTEXT **tctiContext, const char *deviceTctiName )
 {
