+++ test/tpmclient/tpmclient.c
@@ -50,9 +50,9 @@
 
 #include "sapi/tpm20.h"
 #include "sysapi_util.h"
-#include "test/common/sample/sample.h"
+#include "sample.h"
 #include "tpmclient.h"
-#include "common/tcti_util.h"
+#include "tcti_util.h"
 #include "tss2_endian.h"
 
 // This is done to allow the tests to access fields
@@ -65,7 +65,7 @@
 
 #include "tcti/tcti_device.h"
 #include "tcti/tcti_socket.h"
-#include "common/syscontext.h"
+#include "syscontext.h"
 #include "common/debug.h"
 
 //
